% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danish.R
\docType{data}
\name{danish}
\alias{danish}
\title{Danish Fire Insurance Claims}
\format{
A numeric vector containing 2167 observations, with a times attribute which is a POSIXct object of the same length
}
\usage{
danish
}
\description{
These data describe large fire insurance claims in Denmark from Thursday
3rd January 1980 until Monday 31st December 1990. The data are contained in
a numeric vector. The dates of each observation are contained in a times
attribute, which is an object of class "POSIXct" (see  DateTimeClasses). They
were supplied by Mette Rytgaard of Copenhagen Re. Note that these data form
an irregular time series.
}
\keyword{datasets}
