% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpddiag.R
\name{gpd.diag}
\alias{gpd.diag}
\title{Diagnostic plots for a fit to the Generalized Pareto distribution}
\usage{
gpd.diag(z, conf = 0.95, method = c("delta", "profile"), safe = FALSE)
}
\arguments{
\item{z}{A class object returned from `gpdFit'.}

\item{conf}{Confidence level used in the return level plot.}

\item{method}{The method to compute the return level confidence interval - either delta method (default) or
profile likelihood. Choosing profile likelihood may be quite slow.}

\item{safe}{Binary (default FALSE) For when plot margins are too large}
}
\value{
For stationary models, provides return level, density, probability, and quantile plots for the GPD exceedances. The
overlaid density is the `true' density for the estimated parameters. For nonstationary models, provides
residual probability and quantile plots. In addition, nonstationary models provide plots of the residuals vs.
the parameter covariates.
}
\description{
Allows calling of a 2x2 diagnostic plot, different plots available for
stationary vs non-stationary models.
The extension from EVA allows for direct calling to gpd.diag and plots
produced. Additionally, it exposes the individual plots for custom sizing
(especially via .rmd or .rnw)
}
\details{
See the reference for details on how return levels are calculated.
}
\examples{
\dontrun{
x <- rgpd(10000, loc = 0.5, scale = 1, shape = 0.1)
z <- gpd.fit(x, nextremes = 500)
plot(z)
}
}
\references{
Coles, S. (2001). An introduction to statistical modeling of extreme values (Vol. 208). London: Springer.

Modified from "eva : Extreme Value Analysis with Goodness-of-Fit Testing."
}
