% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd.R
\name{gpd.fitu}
\alias{gpd.fitu}
\title{GPD Compute mle over a range of u}
\usage{
gpd.fitu(data, umin, umax, nint = 10, ...)
}
\arguments{
\item{data}{dataset}

\item{umin}{lower xi value}

\item{umax}{upper xi value}

\item{nint}{Number of intervals \code{default} is set to \code{10}}

\item{...}{additional parameters for the gpd fitting function}
}
\description{
GPD Compute mle over a range of u
}
\examples{
data(danish)
gpd.fitu(danish, 0, 20)
}
