% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdplot.R
\name{gpd.plot}
\alias{gpd.plot}
\title{GPD Accessory Plots}
\usage{
gpd.plot(x, pick = 1, optlog = NA, extend = 1.5, labels = TRUE, ...)
}
\arguments{
\item{x}{GPD Object}

\item{pick}{Plot selection type}

\item{optlog}{optional argument for plots 1 and 2 giving a particular choice of logarithmic axes: "x" x-axis only; "y" y-axis only; "xy" both axes; "" neither axis.}

\item{extend}{optional argument for plots 1 and 2 expressing how far x-axis should extend as a multiple of the largest data value. This argument must take values greater than 1 and is useful for showing estimated quantiles beyond data.}

\item{labels}{optional argument for plots 1 and 2 specifying whether or not axes should be labelled}

\item{...}{other graphics parameters}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
The plot method plot.gpd provides four different plots for assessing fitted GPD model.
}
\details{
1 ~ Excess Distribution
2 ~ Tail of Underlying Distribution
3 ~ Scatterplot of Residuals
4 ~ QQplot of Residuals
}
\examples{
data(danish)
out = gpd.fit(danish, 10)
gpd.plot(out, pick=1)
}
