% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpddiag.R
\name{gpd.pp}
\alias{gpd.pp}
\title{PP-Probability Plot}
\usage{
gpd.pp(z, scalevec, shapevec)
}
\arguments{
\item{z}{GPD Model}

\item{scalevec}{Vector of scale}

\item{shapevec}{Vector of Shape}
}
\description{
PP-Probability Plot
}
