% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdprof.R
\name{gpd.prof}
\alias{gpd.prof}
\title{GPD Profile Log Likelihood for m year return level}
\usage{
gpd.prof(z, m, xlow, xup, npy = 365, conf = 0.95, nint = 100)
}
\arguments{
\item{z}{dataset}

\item{m}{number of years}

\item{xlow}{lower xi value}

\item{xup}{upper xi value}

\item{npy}{number of days in year \code{default} is set to \code{365}}

\item{conf}{Confidence size \code{default} is set to \code{0.95}}

\item{nint}{Number of intervals \code{default} is set to \code{100}}
}
\description{
Creates m year return level plot for profile log liklihood
}
\examples{
data(danish)
out = gpd.fit(danish, 10)
gpd.prof(out, m = 10, 55, 75)
}
