% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdprof.R
\name{gpd.profxi}
\alias{gpd.profxi}
\title{GPD Profile Log Likelihood for \eqn{\xi}}
\usage{
gpd.profxi(z, xlow, xup, conf = 0.95, nint = 100)
}
\arguments{
\item{z}{dataset}

\item{xlow}{lower \eqn{\xi} value}

\item{xup}{upper \eqn{\xi} value}

\item{conf}{Confidence size \code{default} is set to \code{0.95}}

\item{nint}{Number of intervals \code{default} is set to \code{100}}
}
\description{
Creates \eqn{\xi} profile log liklihood
}
\examples{
data(danish)
out = gpd.fit(danish, 10)
gpd.profxi(out, -0.1, 0.2)
}
