% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpddiag.R
\name{gpd.qq2}
\alias{gpd.qq2}
\title{QQ-Probability Plot}
\usage{
gpd.qq2(z, scalevec, shapevec)
}
\arguments{
\item{z}{GPD Model}

\item{scalevec}{Vector of scale}

\item{shapevec}{Vector of Shape}
}
\description{
QQ-Probability Plot
}
