% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpddiag.R
\name{gpd.resid}
\alias{gpd.resid}
\title{residuals vs. the covariates Plot}
\usage{
gpd.resid(z, scalevec, shapevec)
}
\arguments{
\item{z}{GPD Model}

\item{scalevec}{Vector of scale}

\item{shapevec}{Vector of Shape}
}
\description{
residuals vs. the covariates Plot
}
