% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdrl.R
\name{gpd.rl}
\alias{gpd.rl}
\title{GPD Return Level Estimate and Confidence Interval for Stationary Models}
\usage{
gpd.rl(
  z,
  period,
  conf = 0.95,
  method = c("delta", "profile"),
  plot = TRUE,
  opt = c("Nelder-Mead")
)
}
\arguments{
\item{z}{An object of class `gpd.fit'.}

\item{period}{The number of periods to use for the return level.}

\item{conf}{Confidence level. Defaults to 95 percent.}

\item{method}{The method to compute the confidence interval - either delta method (default) or profile likelihood.}

\item{plot}{Plot the profile likelihood and estimate (vertical line)?}

\item{opt}{Optimization method to maximize the profile likelihood if that is selected. Argument passed to optim. The
default method is Nelder-Mead.}
}
\value{
\item{Estimate}{Estimated m-period return level.}
\item{CI}{Confidence interval for the m-period return level.}
\item{Period}{The period length used.}
\item{ConfLevel}{The confidence level used.}
}
\description{
Computes stationary m-period return level estimate and interval for the Generalized Pareto distribution,
using either the delta method or profile likelihood.
}
\details{
WARNING : method = profile currently does not work and will error

Caution: The profile likelihood optimization may be slow for large datasets.
}
\examples{
x = rgpd(5000, loc = 0, scale = 1, shape = -0.1)
## Compute 50-period return level.
z = gpd.fit(x, nextremes = 200)
gpd.rl(z, period = 50, method = "delta")
gpd.rl(z, period = 50, method = "delta")
}
\references{
Coles, S. (2001). An introduction to statistical modeling of extreme values (Vol. 208). London: Springer.
}
