% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpddiag.R
\name{gpd.rlPlot}
\alias{gpd.rlPlot}
\title{Return level Plot}
\usage{
gpd.rlPlot(z, conf = 0.95, method = c("delta", "profile"))
}
\arguments{
\item{z}{GPD Model}

\item{conf}{Confidence Level}

\item{method}{Method of Optimisation type}
}
\description{
Return level Plot
}
